WIP: support set pull interval push delay for each repo and app pair: add `AutomationUtil.getAppAndRepoSpecifiedSettingsActuallyBeUsed()` to get pull/push settings which actually be used

+++ app/src/main/java/com/catpuppyapp/puppygit/settings/PackageNameAndRepoSettings.kt
@@ -1,6 +1,6 @@
 package com.catpuppyapp.puppygit.settings
 
-import com.catpuppyapp.puppygit.utils.parseIntOrDefault
+import com.catpuppyapp.puppygit.utils.parseLongOrDefault
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -10,12 +10,12 @@ data class PackageNameAndRepoSettings(
 ) {
 
     // get valid number or empty string, never return invalid number and non-empty string
-    fun getPullIntervalFormatted() = parseIntOrDefault(pullInterval, null)?.toString() ?: ""
-    fun getPushDelayFormatted() = parseIntOrDefault(pushDelay, null)?.toString() ?: ""
+    fun getPullIntervalFormatted() = parseLongOrDefault(pullInterval, null)?.toString() ?: ""
+    fun getPushDelayFormatted() = parseLongOrDefault(pushDelay, null)?.toString() ?: ""
 
     companion object {
         // if is not a valid number, save empty string to settings
-        fun formatPullIntervalBeforeSaving(value:String) = parseIntOrDefault(value, null)?.toString() ?: ""
-        fun formatPushDelayBeforeSaving(value:String) = parseIntOrDefault(value, null)?.toString() ?: ""
+        fun formatPullIntervalBeforeSaving(value:String) = parseLongOrDefault(value, null)?.toString() ?: ""
+        fun formatPushDelayBeforeSaving(value:String) = parseLongOrDefault(value, null)?.toString() ?: ""
     }
 }
